safari-534.53-branch/LayoutTests/fast/js/function-declarations-in-switch-statement-expected.txt

WARN: shouldBe() expects string arguments
PASS 20 is 20
WARN: shouldBe() expects string arguments
PASS 20 is 20
WARN: shouldBe() expects string arguments
PASS 20 is 20
PASS successfullyParsed is true

TEST COMPLETE
